*: consolidate sockopt_ apis in sockopt.c module
Several sockopt_xxx apis were in sockunion - move them into
the sockopt module and header.
